requirements.txt
django
celery
django-celery-beat
notion-client
djangorestframework
redis